feat: allow peeking for component signals
this also removes the ability to peek signals by passing it to the
untrack method, it will now only accept a callback function to call.

reasoning being that the way it works can be rather confusing,
especially if it turns out that the value isn't a signal despite
expecting that it is.

if the value happens to be a function then it should be returning the
function, not call the function and get its return value.

this is not considered breaking change.